FEAT: Partially add SettingsViewController
In this commit, AccountViewController has been added. Both the UI and
code have been fully set up. The segue from SettingsViewController to
AccountViewController has also be established, but other segues have
not been established because there is nothing to segue into.

There are a lot of changes to code in other View Controllers, mostly
to comments for better readability and understanding. To learn more,
please clone or download the project and compare the differences.

In future commits, AccountViewController will be able to segue into
DeleteViewController, a View Controller where the user can choose to
delete their data or account as a whole.