FEAT: Partially add SettingsViewController
In this commit, DeleteViewController has been added. The UI has
been set up, but the code is still not configured. There are no
segues established because there is nothing to segue into.

A few changes to code and comments were made in this commit,
and a configureButton() has now been changed to include the
colours required for the button.

In future commits, DeleteViewController will be able to segue into
ReauthenticatationViewController to be reauthenticate if any
issues arise.