FEAT: Add SettingsViewController
In this commit, ReauthenticationViewController has been added,
finalising and finishing the contents of the entirety of the
Account section; users will be able to manage their account
details, as well as delete any data that they have and remove
themselves completely as a whole.

In future commits, code may still be better refined, if not
changed completely, wherever the developers of the application
see fit.