STYLE: Reorganise code
In this commit, the code has been greatly updated with better
comments and minor refinements to logic executions of the code.

Future commits may still make changes to the code, wherever the
developers of the application see fit or in any critical event
where the breaking of the application is imminent.