FEAT: Add set up Storyboard
In this commit, the Setup storyboard group has been created and
successfully completed. Multiple bugs found in the application
concerning the storing of the Firestore database is also fixed
in this commit.

The Setup storyboard group asks the students for their levels -
to cater the pre-researched provided value scores - as well as
ask for their Mother Tongue language name. The Setup storyboard
will then be dismissed. However, note that there is a bug that
will make IntroductionViewController appear again after concluding.
This issue is acknowledged and will be fixed in future commits.

Besides adding the Storyboard, multiple changes in the code were
also made; clone or download the repository and the project to learn
more about the changes.