FEAT: Partially complete Subjects Storyboard
This commit partially completes the Subjects Storyboard that is newly
created in this commit. Subjects Storyboard acts as the main location
where most of the user interaction will take place; students will be
able to add in marks for their examinations and see data, statistics,
and graphs in the end result.

Currently, only SubjectsViewController has been established; code has
been implemented to fetch the subjects for the students to view. Future
commits will make changes to the Storyboard as a whole and may improve
the general user experience whenever the developers see fit.

In future commits, SubjectsViewController may be changed, and
DetailViewController will be set up.