FEAT: Import Charts framework
This commit adds and imports the Charts framework. As this commit is
initially meant to be committed together with Commit 18 (FEAT: Partially
complete Subjects Storyboard), it is recommended that this commit is
pulled immediately into the local repositories.

In future commits, more frameworks may be removed or added, wherever the
developers see fit.