FEAT: Refinements to Subjects
This commit establishes the use of the Charts framework to create a custom
graph for the user. This graph verifies if there is substantial information
to chart data before drawing, therefore effectively allowing for good user
UX.

There may be minor refinements in this commit as well, and therefore it is
advised to clone, pull, or download the project to see more changes.